pyproject.toml
[project]
name = "cua-lark"
version = "0.1.0"
description = "基于多模态大模型的飞书桌面端智能操作代理"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.10"
dependencies = [
    "dashscope>=1.25.0",
    "mss>=9.0.1",
    "Pillow>=10.0.0",
    "PyAutoGUI>=0.9.54",
    "pyperclip>=1.8.2",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
]

[project.scripts]
cua-lark = "src.core.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
ocr = [
    "EasyOCR>=1.7.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
namespaces = false

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
